agent/requirements.txt
boto3>=1.42.60
strands-agents[otel]>=1.28.0
bedrock-agentcore>=1.1.1
aws-opentelemetry-distro>=0.14.1
pydantic>=2.12.5
